feat: use most recent post time for source last updated and reduce check sources frequency
It is apparently common enough to see posts with a published time that's in the future, using the most recent post's published time instead of start date helps prevent multiple emails of the same post when that happens. It might also help with not missing posts that are released after a check with a time before the check as long as they are more recent than previous posts.
Decrease the frequency of check sources to cut down on db reads and hopefully solve the issue of missing posts that are published with a time older than the most recent post (as in, give a wider time for stuff like that to happen so we hopefully still get to check said earlier posts. this doesn't actually fix the issue and that can probably only be solved by recording seen posts and checking against that to see which emails to send as opposed to just keeping track of most recent send times)