mail posts in reverse, print number of new posts for changed sources
Print number for convenience
Mail posts in reverse so new posts are more likely to be received later. Emails aren't guaranteed to be recieved in the order sent and most of the time there's only one or two posts, but when there are many posts (when a source is renabled after being disabled for a while for example) it's nice to read the older ones to be sent before the newer ones. Free bonus: acivities for new posts will show up higher on the list of activities so checking the most recent activity for a source is easier